+++ src/main/java/com/toy/getyourfriday/response/UpdateResponse.java
@@ -1,5 +1,6 @@
-package com.toy.getyourfriday.domain.response;
+package com.toy.getyourfriday.response;
 
+import com.toy.getyourfriday.domain.product.Product;
 import com.toy.getyourfriday.domain.user.User;
 import com.toy.getyourfriday.dto.InlineKeyboardMarkup;
 import lombok.Getter;
@@ -12,23 +13,24 @@ import java.net.URI;
 import java.util.Objects;
 
 @Getter
-public class UserResponse extends WebClientResponse {
+public class UpdateResponse extends WebClientResponse {
 
-    private final String message;
+    private final Product updateProduct;
 
-    public UserResponse(Integer chatId, String message) {
+    public UpdateResponse(Integer chatId, Product updateProduct) {
         super(chatId);
-        this.message = message;
+        this.updateProduct = updateProduct;
     }
 
-    public static UserResponse of(User user, String message) {
-        return new UserResponse(user.getChatId(), message);
+    public static UpdateResponse of(User user, Product updateProduct) {
+        return new UpdateResponse(user.getChatId(), updateProduct);
     }
 
     @Override
     public Mono<?> send(WebClient webClient) {
-        return webClient.get()
+        return webClient.post()
                 .uri(this::buildUri)
+                .bodyValue(InlineKeyboardMarkup.from(updateProduct))
                 .retrieve()
                 .onStatus(HttpStatus::isError, WebClientResponse::mapToHttpStatusCodeException)
                 .bodyToMono(InlineKeyboardMarkup.class);
@@ -37,21 +39,21 @@ public class UserResponse extends WebClientResponse {
     @Override
     protected URI buildUri(UriBuilder uriBuilder) {
         return uriBuilder.queryParam("chat_id", chatId)
-                .queryParam("text", message)
+                .queryParam("text", updateProduct.getThumbnail())
                 .build();
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof UserResponse)) return false;
+        if (!(o instanceof UpdateResponse)) return false;
         if (!super.equals(o)) return false;
-        UserResponse that = (UserResponse) o;
-        return getMessage().equals(that.getMessage());
+        UpdateResponse that = (UpdateResponse) o;
+        return getUpdateProduct().equals(that.getUpdateProduct());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), getMessage());
+        return Objects.hash(super.hashCode(), getUpdateProduct());
     }
 }
