+++ src/main/java/com/toy/getyourfriday/response/UserResponse.java
@@ -1,6 +1,5 @@
-package com.toy.getyourfriday.domain.response;
+package com.toy.getyourfriday.response;
 
-import com.toy.getyourfriday.domain.product.Product;
 import com.toy.getyourfriday.domain.user.User;
 import com.toy.getyourfriday.dto.InlineKeyboardMarkup;
 import lombok.Getter;
@@ -13,24 +12,23 @@ import java.net.URI;
 import java.util.Objects;
 
 @Getter
-public class UpdateResponse extends WebClientResponse {
+public class UserResponse extends WebClientResponse {
 
-    private final Product updateProduct;
+    private final String message;
 
-    public UpdateResponse(Integer chatId, Product updateProduct) {
+    public UserResponse(Integer chatId, String message) {
         super(chatId);
-        this.updateProduct = updateProduct;
+        this.message = message;
     }
 
-    public static UpdateResponse of(User user, Product updateProduct) {
-        return new UpdateResponse(user.getChatId(), updateProduct);
+    public static UserResponse of(User user, String message) {
+        return new UserResponse(user.getChatId(), message);
     }
 
     @Override
     public Mono<?> send(WebClient webClient) {
-        return webClient.post()
+        return webClient.get()
                 .uri(this::buildUri)
-                .bodyValue(InlineKeyboardMarkup.from(updateProduct))
                 .retrieve()
                 .onStatus(HttpStatus::isError, WebClientResponse::mapToHttpStatusCodeException)
                 .bodyToMono(InlineKeyboardMarkup.class);
@@ -39,21 +37,21 @@ public class UpdateResponse extends WebClientResponse {
     @Override
     protected URI buildUri(UriBuilder uriBuilder) {
         return uriBuilder.queryParam("chat_id", chatId)
-                .queryParam("text", updateProduct.getThumbnail())
+                .queryParam("text", message)
                 .build();
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof UpdateResponse)) return false;
+        if (!(o instanceof UserResponse)) return false;
         if (!super.equals(o)) return false;
-        UpdateResponse that = (UpdateResponse) o;
-        return getUpdateProduct().equals(that.getUpdateProduct());
+        UserResponse that = (UserResponse) o;
+        return getMessage().equals(that.getMessage());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), getUpdateProduct());
+        return Objects.hash(super.hashCode(), getMessage());
     }
 }
