+++ src/main/java/com/toy/getyourfriday/response/WebClientResponse.java
@@ -1,4 +1,4 @@
-package com.toy.getyourfriday.domain.response;
+package com.toy.getyourfriday.response;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
