+++ src/main/java/com/toy/getyourfriday/service/ScrapingService.java
@@ -1,8 +1,8 @@
 package com.toy.getyourfriday.service;
 
-import com.toy.getyourfriday.domain.response.WebClientResponse;
 import com.toy.getyourfriday.dto.RegisterRequest;
 import com.toy.getyourfriday.dto.RemoveRequest;
+import com.toy.getyourfriday.response.WebClientResponse;
 
 public interface ScrapingService {
