+++ src/test/java/com/toy/getyourfriday/service/UpdateServiceTest.java
@@ -3,10 +3,10 @@ package com.toy.getyourfriday.service;
 import com.toy.getyourfriday.component.ModelUrlParser;
 import com.toy.getyourfriday.domain.product.Product;
 import com.toy.getyourfriday.domain.product.Products;
-import com.toy.getyourfriday.domain.response.UpdateResponse;
 import com.toy.getyourfriday.domain.scraping.ModelUrl;
 import com.toy.getyourfriday.domain.user.User;
 import com.toy.getyourfriday.domain.user.UserRepository;
+import com.toy.getyourfriday.response.UpdateResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
